chore: address gosec G115 issue

+++ pkg/homekit/colortemperature_light.go
@@ -75,7 +75,7 @@ func createColorTemperatureLightAccessory(light *hue.Light, bridge hue.Bridger) 
 
 	// create the lightbulb accessory
 	ac := NewColorTemperatureLight(accessory.Info{
-		ID:               uint64(id + 1),
+		ID:               uint64(id + 1), // #nosec G115 IDs will always be smaller
 		Name:             light.Name,
 		Model:            light.ModelID,
 		Manufacturer:     light.ManufacturerName,
@@ -93,7 +93,7 @@ func createColorTemperatureLightAccessory(light *hue.Light, bridge hue.Bridger) 
 		err := bridge.LightUpdateState(light, &hue.State{On: on})
 
 		log.WithFields(log.Fields{
-			"id":   uint64(id + 1),
+			"id":   id + 1,
 			"name": light.Name,
 			"type": light.Type,
 		}).Debugf("trigger state: %t", on)
@@ -102,7 +102,7 @@ func createColorTemperatureLightAccessory(light *hue.Light, bridge hue.Bridger) 
 		if err != nil {
 			// ...log it
 			log.WithFields(log.Fields{
-				"id":    uint64(id + 1),
+				"id":    id + 1,
 				"name":  light.Name,
 				"state": on,
 				"on":    "on",
@@ -139,7 +139,7 @@ func createColorTemperatureLightAccessory(light *hue.Light, bridge hue.Bridger) 
 		err := bridge.LightUpdateState(light, &hue.State{On: true, Brightness: bri})
 
 		log.WithFields(log.Fields{
-			"id":   uint64(id + 1),
+			"id":   id + 1,
 			"name": light.Name,
 			"type": light.Type,
 		}).Debugf("change brightness: %d", bri)
@@ -148,7 +148,7 @@ func createColorTemperatureLightAccessory(light *hue.Light, bridge hue.Bridger) 
 		if err != nil {
 			// ...log it
 			log.WithFields(log.Fields{
-				"id":   uint64(id + 1),
+				"id":   id + 1,
 				"name": light.Name,
 				"bri":  bri,
 				"on":   "brightness",
@@ -186,7 +186,7 @@ func createColorTemperatureLightAccessory(light *hue.Light, bridge hue.Bridger) 
 		err := bridge.LightUpdateState(light, &hue.State{On: true, ColorTemperature: colorTemperature})
 
 		log.WithFields(log.Fields{
-			"id":   uint64(id + 1),
+			"id":   id + 1,
 			"name": light.Name,
 			"type": light.Type,
 		}).Debugf("change color-temperature: %d", colorTemperature)
@@ -195,7 +195,7 @@ func createColorTemperatureLightAccessory(light *hue.Light, bridge hue.Bridger) 
 		if err != nil {
 			// ...log it
 			log.WithFields(log.Fields{
-				"id":               uint64(id + 1),
+				"id":               id + 1,
 				"name":             light.Name,
 				"colorTemperature": colorTemperature,
 				"on":               "color-temperature",
